Added item drops

+++ IdleRpg/Game/Core/ItemDrop.cs
@@ -3,5 +3,5 @@
 public class ItemDrop
 {
     public required Enum Item { get; init; }
-    public required int DropChance { get; init; }
+    public required double DropChance { get; init; }
 }
